feat: add dynamic page titles (browser tab metadata) (#250)
