+++ src/app/(app)/[workspaceSlug]/settings/members/page.tsx
@@ -1,7 +1,28 @@
+import type { Metadata } from "next";
 import { notFound, redirect } from "next/navigation";
 import { createClient } from "@/lib/supabase/server";
 import { MembersPage } from "@/components/members/members-page";
 import type { MemberWithProfile, WorkspaceInviteWithInviter } from "@/lib/types";
+
+export async function generateMetadata({
+  params,
+}: {
+  params: Promise<{ workspaceSlug: string }>;
+}): Promise<Metadata> {
+  const { workspaceSlug } = await params;
+  const supabase = await createClient();
+
+  const { data: workspace } = await supabase
+    .from("workspaces")
+    .select("name")
+    .eq("slug", workspaceSlug)
+    .maybeSingle();
+
+  const workspaceName = workspace?.name ?? workspaceSlug;
+  return {
+    title: `Members — ${workspaceName}`,
+  };
+}
 
 export default async function WorkspaceMembersPage({
   params,
