+++ src/app/(auth)/sign-in/page.tsx
@@ -1,140 +1,10 @@
-"use client";
+import type { Metadata } from "next";
+import { SignInForm } from "./sign-in-form";
 
-import { Suspense, useState } from "react";
-import { useRouter, useSearchParams } from "next/navigation";
-import Link from "next/link";
-import { createClient } from "@/lib/supabase/client";
-import { Button } from "@/components/ui/button";
-import { Input } from "@/components/ui/input";
-import { Label } from "@/components/ui/label";
-import {
-  Card,
-  CardContent,
-  CardDescription,
-  CardHeader,
-  CardTitle,
-} from "@/components/ui/card";
-import { OAuthButtons } from "@/components/auth/oauth-buttons";
-
-function SignInForm() {
-  const router = useRouter();
-  const searchParams = useSearchParams();
-  const confirmed = searchParams.get("confirmed") === "true";
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-  const [error, setError] = useState<string | null>(null);
-  const [loading, setLoading] = useState(false);
-
-  async function handleSubmit(e: React.FormEvent<HTMLFormElement>) {
-    e.preventDefault();
-    setError(null);
-    setLoading(true);
-
-    const supabase = createClient();
-    const { error: signInError } = await supabase.auth.signInWithPassword({
-      email,
-      password,
-    });
-
-    if (signInError) {
-      setError(signInError.message);
-      setLoading(false);
-      return;
-    }
-
-    // Fetch the user's personal workspace to redirect to
-    const {
-      data: { user },
-    } = await supabase.auth.getUser();
-    if (user) {
-      const { data: membership } = await supabase
-        .from("members")
-        .select("workspace_id, workspaces(slug)")
-        .eq("user_id", user.id)
-        .limit(1)
-        .maybeSingle();
-
-      if (membership?.workspaces) {
-        const ws = membership.workspaces as unknown as { slug: string };
-        router.push(`/${ws.slug}`);
-        return;
-      }
-    }
-
-    // Fallback: redirect to root which will handle routing
-    router.push("/");
-  }
-
-  return (
-    <Card>
-      <CardHeader>
-        <CardTitle className="text-2xl font-semibold">Sign in to Memo</CardTitle>
-        <CardDescription>
-          Enter your email and password to continue.
-        </CardDescription>
-      </CardHeader>
-      <CardContent>
-        <form onSubmit={handleSubmit} className="flex flex-col gap-3">
-          <div className="flex flex-col gap-1.5">
-            <Label htmlFor="email">Email</Label>
-            <Input
-              id="email"
-              type="email"
-              placeholder="you@example.com"
-              value={email}
-              onChange={(e) => setEmail(e.target.value)}
-              required
-              autoComplete="email"
-              autoFocus
-            />
-          </div>
-          <div className="flex flex-col gap-1.5">
-            <Label htmlFor="password">Password</Label>
-            <Input
-              id="password"
-              type="password"
-              placeholder="••••••••"
-              value={password}
-              onChange={(e) => setPassword(e.target.value)}
-              required
-              autoComplete="current-password"
-              minLength={6}
-            />
-          </div>
-          {confirmed && (
-            <p className="text-xs text-accent">
-              Email confirmed — you can now sign in.
-            </p>
-          )}
-          {error && <p className="text-xs text-destructive">{error}</p>}
-          <Button type="submit" disabled={loading} className="mt-1">
-            {loading ? "Signing in…" : "Sign in"}
-          </Button>
-        </form>
-        <div className="relative my-4">
-          <div className="absolute inset-0 flex items-center">
-            <span className="w-full border-t border-white/[0.06]" />
-          </div>
-          <div className="relative flex justify-center text-xs">
-            <span className="bg-card px-2 text-muted-foreground">or</span>
-          </div>
-        </div>
-        <OAuthButtons />
-        <p className="mt-4 text-center text-xs text-muted-foreground">
-          Don&apos;t have an account?{" "}
-          <Link href="/sign-up" className="text-accent underline-offset-4 hover:underline">
-            Sign up
-          </Link>
-        </p>
-      </CardContent>
-    </Card>
-  );
-}
+export const metadata: Metadata = {
+  title: "Sign in",
+};
 
 export default function SignInPage() {
-  return (
-    <Suspense>
-      <SignInForm />
-    </Suspense>
-  );
+  return <SignInForm />;
 }
